:hammer: unify access to R2
