+++ lib/catalog/owid/catalog/s3_utils.py
@@ -1,11 +1,16 @@
+import configparser
 import logging
+import os
+import threading
+from functools import lru_cache
 from os import environ as env
-from typing import Any, Tuple
+from typing import Dict, Tuple
 from urllib.parse import urlparse
 
+from botocore.client import BaseClient
 from botocore.exceptions import ClientError
 
-R2_ENDPOINT = "https://078fcdfed9955087315dd86792e71a7e.r2.cloudflarestorage.com"
+BOTO3_CLIENT_LOCK = threading.Lock()
 
 
 def s3_bucket_key(url: str) -> Tuple[str, str]:
@@ -23,7 +28,7 @@ def s3_bucket_key(url: str) -> Tuple[str, str]:
 
 def download(s3_url: str, filename: str, quiet: bool = False) -> None:
     """Download the file at the S3 URL to the given local filename."""
-    client = connect()
+    client = connect_r2()
 
     bucket, key = s3_bucket_key(s3_url)
 
@@ -39,7 +44,7 @@ def download(s3_url: str, filename: str, quiet: bool = False) -> None:
 
 def upload(s3_url: str, filename: str, public: bool = False, quiet: bool = False) -> None:
     """Upload the file at the given local filename to the S3 URL."""
-    client = connect()
+    client = connect_r2()
     bucket, key = s3_bucket_key(s3_url)
     extra_args = {"ACL": "public-read"} if public else {}
     try:
@@ -52,24 +57,60 @@ def upload(s3_url: str, filename: str, public: bool = False, quiet: bool = False
         logging.info("UPLOADED", f"{filename} -> {s3_url}")
 
 
-def connect() -> Any:
+# if R2_ACCESS_KEY and R2_SECRET_KEY are null, try using credentials from rclone config
+def _read_owid_rclone_config() -> Dict[str, str]:
+    # Create a ConfigParser object
+    config = configparser.ConfigParser()
+
+    # Read the configuration file
+    config.read(os.path.expanduser("~/.config/rclone/rclone.conf"))
+
+    return dict(config["owid-r2"].items())
+
+
+def connect_r2() -> BaseClient:
     "Return a connection to Cloudflare's R2."
     import boto3
 
+    # first, get the R2 credentials from dotenv
     R2_ACCESS_KEY = env.get("R2_ACCESS_KEY")
     R2_SECRET_KEY = env.get("R2_SECRET_KEY")
+    R2_ENDPOINT = env.get("R2_ENDPOINT")
+    R2_REGION_NAME = env.get("R2_REGION_NAME")
 
-    assert R2_ACCESS_KEY and R2_SECRET_KEY, "Missing R2_ACCESS_KEY and R2_SECRET_KEY environment variables."
+    # alternatively, get them from rclone config
+    if not R2_ACCESS_KEY or not R2_SECRET_KEY or not R2_ENDPOINT:
+        try:
+            rclone_config = _read_owid_rclone_config()
+            R2_ACCESS_KEY = R2_ACCESS_KEY or rclone_config.get("access_key_id")
+            R2_SECRET_KEY = R2_SECRET_KEY or rclone_config.get("secret_access_key")
+            R2_ENDPOINT = R2_ENDPOINT or rclone_config.get("endpoint")
+            R2_REGION_NAME = R2_REGION_NAME or rclone_config.get("region")
+        except KeyError:
+            pass
 
     client = boto3.client(
         service_name="s3",
-        endpoint_url=R2_ENDPOINT,
         aws_access_key_id=R2_ACCESS_KEY,
         aws_secret_access_key=R2_SECRET_KEY,
-        region_name="auto",
+        endpoint_url=R2_ENDPOINT or "https://078fcdfed9955087315dd86792e71a7e.r2.cloudflarestorage.com",
+        region_name=R2_REGION_NAME or "auto",
     )
 
     return client
+
+
+@lru_cache(maxsize=None)
+def _connect_r2_cached() -> BaseClient:
+    return connect_r2()
+
+
+def connect_r2_cached() -> BaseClient:
+    """Connect to R2, but cache the connection for subsequent calls. This is more efficient than
+    creating a new connection for every request."""
+    # creating a client is not thread safe, lock it
+    with BOTO3_CLIENT_LOCK:
+        return _connect_r2_cached()
 
 
 class MissingCredentialsError(Exception):
